requirements.txt
# ─────────────────────────────────────────────────────────────────────────
# RL Dissertation — Python Dependencies
# ─────────────────────────────────────────────────────────────────────────
# DO NOT install this file directly with pip.
# Run setup_env.py — it picks the right PyTorch build for your hardware.
#
#   python setup_env.py
#
# ─────────────────────────────────────────────────────────────────────────

# Atari environment
gymnasium[atari]>=0.29.0
ale-py>=0.8.0
AutoROM[accept-rom-license]>=0.6.1

# Numerics
numpy>=1.24.0,<2.0.0
pandas>=2.0.0

# Visualisation & analysis
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# Experiment tracking
tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
